Major update to Calculator App
While the original purpose of this calculator app was to get in depth practice with auto layout and stack views, I began to implement actual functionality to the calculator. It still has several bugs and a few operator buttons have not been implemented yet, but it can perform basic addition operations. I'm currently dealing with a bug that overwrites the first operand entered into the list of operands - for example, pressing 9 + 1 yields an answer of 1 instead of 10, as the one is overwriting the 9. Also still need to implement subtraction, multiply and divide button functionality - also working on an algorithm for the calculator to processes concatonated operations while also being mindful of order of operations.